% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{lz4}
\alias{lz4}
\alias{lz4-package}
\title{The lz4 compression library.}
\description{
The \code{lz4} package defines a \code{memCompress}-like interface
to the lz4 compression library written by Yann Collet. lz4 generally
favors speed, especially de-compression speed, over compression.
Use lz4 in speed-sensitive situations that can tolerate slightly
larger compressed objects than available with gzip.
}
\seealso{
\code{\link{lzCompress}}, \code{\link{lzDecompress}}
}

