% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lz4_version}
\alias{lz4_version}
\title{Return the lz4 library version number}
\usage{
lz4_version()
}
\value{
The integer value corresponding to the lz4 library version number
embedded in the R package.
}
\description{
Return the lz4 library version number
}
\examples{
lz4_version()
}

